project 6.8

+++ Project System/PKanri.aspx.cs
@@ -75,13 +75,44 @@ namespace WhereEver.Project_System
                 }
             }
         }
+
+        public string bigname = "";
         protected void DgPKanri_ItemDataBound(object sender, DataGridItemEventArgs e)
         {
             if ((e.Item.ItemType == ListItemType.Item) || (e.Item.ItemType == ListItemType.AlternatingItem))
             {
                 DATASET.DataSet.T_PdbKanriRow dr = (e.Item.DataItem as DataRowView).Row as DATASET.DataSet.T_PdbKanriRow;
-                e.Item.Cells[0].Text = dr.PBigname.ToString();
+                
+                if (dr.PBigname != bigname)
+                {
+                    e.Item.Cells[0].Text = "<font color = white>"+dr.PBigname.ToString();
+                    bigname = dr.PBigname;
+                    if (dr.PBigid!=1)
+                    {
+                        Button uebig = e.Item.FindControl("uebig") as Button;
+                        uebig.Visible = true;
+                    }
+                    else if(dr.PBigid != Insert.GetMaxPBigidRow(Global.GetConnection(), SessionManager.project.PdbRow.Pid).PBigid)
+                    {
+                        Button sitabig = e.Item.FindControl("sitabig") as Button;
+                        sitabig.Visible = true;
+                    }
+                }
+                else
+                {
+                    e.Item.Cells[0].Text = "<font color = black>"+dr.PBigname.ToString();
+                }
                 e.Item.Cells[2].Text = dr.PMiddleid.ToString();
+                if (dr.PMiddleid != 1)
+                {
+                    Button uemiddle = e.Item.FindControl("uemiddle") as Button;
+                    uemiddle.Visible = true;
+                }
+                else if (dr.PMiddleid != Insert.GetMaxPMiddleidRow(Global.GetConnection(), dr.PBigname).PMiddleid)
+                {
+                    Button sitamiddle = e.Item.FindControl("sitamiddle") as Button;
+                    sitamiddle.Visible = true;
+                }
                 e.Item.Cells[3].Text = dr.PMiddlename.ToString();
                 e.Item.Cells[5].Text = dr.PMiddlestart.ToShortDateString();
                 e.Item.Cells[6].Text = dr.PMiddleover.ToShortDateString();
@@ -259,30 +290,42 @@ namespace WhereEver.Project_System
 
         protected void DgPKanri_ItemCommand(object source, DataGridCommandEventArgs e)
         {
-            string bigname = e.Item.Cells[0].Text;
-            string Middleid = e.Item.Cells[2].Text;
-            switch (e.CommandName)
+            if (e.CommandName== "Delete"|| e.CommandName == "uebig" || e.CommandName == "sitabig" || e.CommandName == "uemiddle" || e.CommandName == "sitamiddle")
             {
-                case "Delete":
-                    Delete.DeleteMiddle(bigname, Middleid, SessionManager.project.PdbRow.Pid);
-                    break;
-                case "uebig":
-                    Narabi.uebig(Global.GetConnection(), bigname, SessionManager.project.PdbRow.Pid);
-                    CreateDataGrid(SessionManager.project.PdbRow.Pid);
-                    break;
-                case "sitabig":
-                    break;
-                case "uemiddle":
-                    break;
-                case "sitamiddle":
-                    break;
-                default:
-                    // Do nothing.
-                    break;
+                string name = e.Item.Cells[0].Text;
+                string[] arr = name.Split('>');
+                string bigname = arr[1];
+                string Middleid = e.Item.Cells[2].Text;
+                switch (e.CommandName)
+                {
+                    case "Delete":
+                        Delete.DeleteMiddle(bigname, Middleid, SessionManager.project.PdbRow.Pid);
+                        break;
+                    case "uebig":
+                        Narabi.uebig(Global.GetConnection(), bigname, SessionManager.project.PdbRow.Pid);
+                        CreateDataGrid(SessionManager.project.PdbRow.Pid);
+                        break;
+                    case "sitabig":
+                        Narabi.sitabig(Global.GetConnection(), bigname, SessionManager.project.PdbRow.Pid);
+                        CreateDataGrid(SessionManager.project.PdbRow.Pid);
+                        break;
+                    case "uemiddle":
+                        Narabi.uemiddle(Global.GetConnection(), Middleid, bigname, SessionManager.project.PdbRow.Pid);
+                        CreateDataGrid(SessionManager.project.PdbRow.Pid);
+                        break;
+                    case "sitamiddle":
+                        Narabi.sitamiddle(Global.GetConnection(), Middleid, bigname, SessionManager.project.PdbRow.Pid);
+                        CreateDataGrid(SessionManager.project.PdbRow.Pid);
+                        break;
+                    default:
+                        // Do nothing.
+                        break;
 
+                }
+                DgPKanri.EditItemIndex = -1;
+                CreateDataGrid(SessionManager.project.PdbRow.Pid);
             }
-            DgPKanri.EditItemIndex = -1;
-            CreateDataGrid(SessionManager.project.PdbRow.Pid);
+                
         }
 
         protected void btnDeleteBig_Click(object sender, EventArgs e)
